Update productDetail, cart

+++ src/main/java/com/poly/Entity/Cart_Items.java
@@ -34,7 +34,7 @@ public class Cart_Items implements Serializable {
     private Carts Carts;
 
     @Column(name = "quantity")
-    private Float Quantity;
+    private int Quantity;
 
     @Column(name = "price")
     private Float Price;
